Refact: Renamed Components for consistency
