+++ components/NotesContainer.tsx
