+++ components/VideoCard.tsx
