+++ pages/history.tsx
@@ -4,8 +4,8 @@ import { NextPage } from "next";
 import { addVideoToHistory, getVideoHistory } from "../redux/videoHistorySlice";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../redux/store";
-import VideoCard from "../components/videoCard";
-import Navbar from "../components/navbar";
+import VideoCard from "../components/VideoCard";
+import Navbar from "../components/Navbar";
 
 const History: NextPage = () => {
   const userState = useSelector((state: RootState) => state.auth.user);
