+++ pages/login.tsx
@@ -1,7 +1,7 @@
 import dynamic from "next/dynamic";
 import React from "react";
 
-const Login = dynamic(() => import("../components/login"), {
+const Login = dynamic(() => import("../components/Login"), {
   ssr: false,
 });
